Add separate view model for movie details page.
