+++ app/src/main/java/ru/sandbox/androidacademyapp/MoviesViewModelFactory.kt
@@ -8,6 +8,7 @@ import ru.sandbox.androidacademyapp.repository.MovieRepository
 class MoviesViewModelFactory() : ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T = when (modelClass) {
         MoviesViewModel::class.java -> MoviesViewModel(MovieRepository(MoviesApi.create()))
+        MovieDetailsViewModel::class.java -> MovieDetailsViewModel(MovieRepository(MoviesApi.create()))
         else -> throw IllegalArgumentException("$modelClass is not registered ViewModel")
     } as T
 }
